fix trivial size_t warnings

+++ onnxruntime/core/optimizer/slice_elimination.cc
@@ -41,7 +41,7 @@ bool EliminateSlice::SatisfyCondition(const Graph& graph, const Node& node) {
 
   // For now eliminate slice operators if starts=0 and ends=MAX_INT or -1.
   // TODO: Take into account the input's shape to get a tighter bound for the ends.
-  for (int i = 0; i < axes.size(); ++i) {
+  for (size_t i = 0; i < axes.size(); ++i) {
     if (starts[i] > 0 || starts[i] < 0 ||
         (ends[i] > 0 && ends[i] < INT64_MAX)) {
       return false;
